011022.1

+++ PullUpsDapper/Repository.cs
@@ -1,5 +1,4 @@
-﻿
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -25,7 +24,7 @@ namespace PullUpsDapper
     }
     public class UserRepository : IUser
     {
-        public string ConnString { get; set; }
+        public string? ConnString { get; set; }
         public List<User> GetUsers()
         {
             ConnString = DBConnection.ConnectionString();
@@ -37,9 +36,12 @@ namespace PullUpsDapper
         {
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
-            string lvl = conn.ExecuteScalar<string>("SELECT users.level FROM  pulls.users  WHERE users.user_id = " + userId + ";");
-            int count = conn.ExecuteScalar<int>("SELECT count(*) FROM  pulls.users  WHERE users.user_id = " + userId + ";");
-            //bool program = conn.ExecuteScalar<bool>("SELECT count(*) FROM  pulls.lvl_user_program  WHERE lvl_user_program.user_id = " + userId + ";");
+            //string lvl = conn.ExecuteScalar<string>("SELECT users.level FROM  pulls.users  WHERE users.user_id = " + userId + ";");
+            //int count = conn.ExecuteScalar<int>("SELECT count(*) FROM  pulls.users  WHERE users.user_id = " + userId + ";");
+
+            string lvl = conn.ExecuteScalar<string>("SELECT users.level FROM  pulls.users  WHERE users.user_id = @users.user_id;", new { user_id = userId });
+            int count = conn.ExecuteScalar<int>("SELECT count(*) FROM  pulls.users  WHERE users.user_id = @users.user_id;", new { user_id = userId });
+
             conn.Close();
             return (lvl, count);
         }
@@ -49,59 +51,41 @@ namespace PullUpsDapper
             // для теста
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
-            string query = "SELECT users.level FROM  pulls.users  WHERE users.user_id = @users.user_id;";
+            string query = @"SELECT users.level FROM pulls.users  WHERE users.user_id = @users.user_id;";
             var lvl = conn.QueryFirstOrDefault<User>(query, new { user_id = userId });
 
             return lvl;
         }
 
-
-
         public void CreateUser(User user)
         {
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
-            var sqlQuery = "INSERT INTO pulls.users  (user_id, name) VALUES ('" + user.IdUser + "', '" + user.Name + "')";
-            conn.Execute(sqlQuery);
+            var sqlQuery = @"INSERT INTO pulls.users (user_id, name) VALUES (@user_id, @name)";
+            conn.Execute(sqlQuery, new { @user_id = user.IdUser, @name = user.Name });
             conn.Close();
         }
 
         public void UpdateUser(string lvl, long userId)
         {
+            //cnn.Execute("update Table set val = @val where Id = @id", new { val, id = 1 });
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
-            var sqlQuery = "UPDATE pulls.users  SET level = '" + lvl + "' WHERE users.user_id = " + userId + ";";
-            conn.Execute(sqlQuery);
+            var sqlQuery = @"UPDATE pulls.users SET level = '" + lvl + "' WHERE @users.user_id = @user_id;";
+            conn.Execute(sqlQuery, new { lvl, @user_id = userId });
             conn.Close();
         }
 
         public void CreateTrainingProgram(string lvl, long userId)
         {
             ConnString = DBConnection.ConnectionString();
-            // var (result, UserProgram) = CreateProgram.CreateUserProgram(lvl, userId);
             var result = CreateProgram.CreateUserProgram(lvl, userId);
             using var conn = new NpgsqlConnection(ConnString);
-            //for (int i = 0; i < UserProgram.Count; i++)
-            //{  
-            //    TrainingProgram program = UserProgram[i];
-            //    var sqlQuery = "INSERT INTO pulls.user_program (user_id, week, approach, pulls)  VALUES ('"
-            //        + program.Id + "', '"
-            //        + program.Week + "', '"
-            //        + program.Approach + "', '"
-            //        + program.Pulls + "'"
-            //        + ")";
-            //    conn.Execute(sqlQuery); 
-            //}
             for (int i = 0; i < result.Count; i++)
             {
                 DayResult res = result[i];
-                var sqlQuery = "INSERT INTO pulls.day_result (user_id, week, date, pulls)  VALUES ('"
-                    + res.Id + "', '"
-                    + res.Week + "', '"
-                    + res.Date + "', '"
-                    + 0 + "'"
-                    + ")";
-                conn.Execute(sqlQuery);
+                var sqlQuery = @"INSERT INTO pulls.day_result (user_id, week, date, pulls) VALUES (@user_id, @week, @date, @pulls)";
+                conn.Execute(sqlQuery, new { @user_id = res.Id, @week = res.Week, @date = res.Date, @pulls = 0 });
             }
             conn.Close();
         }
@@ -111,11 +95,13 @@ namespace PullUpsDapper
             ConnString = DBConnection.ConnectionString();
             var date = DateTime.Now;
             using var conn = new NpgsqlConnection(ConnString);
-            string sqlQuery = "SELECT a.approach , a.pulls" +
-                              " FROM  pulls.lvl_user_program a LEFT JOIN pulls.day_result b ON a.week = b.week " +
-                              " WHERE a.level = (Select level From pulls.users Where user_id = " + userId + ")::text  and b.date = CAST('" + date + "' as Date);";
 
-            var dayProgram = conn.Query<UserDayProgram>(sqlQuery);
+            string sqlQuery = "SELECT a.approach , a.pulls" +
+                  " FROM  pulls.lvl_user_program a LEFT JOIN pulls.day_result b ON a.week = b.week " +
+                  " WHERE a.level = (Select level From pulls.users Where user_id = @users.user_id)::text  and b.date = CAST(@date as Date);";
+
+            var dayProgram = conn.Query<UserDayProgram>(sqlQuery, new { @user_id = userId, @date = date });
+
             List<UserDayProgram> userDayProgram = new List<UserDayProgram>();
 
             foreach (var item in dayProgram)
@@ -133,10 +119,10 @@ namespace PullUpsDapper
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
             string sqlQuery;
-            sqlQuery = "Select sum(pulls) From pulls.lvl_user_program WHERE " +
-                              "week = (Select week From pulls.day_result WHERE date = CAST('" + date + "' as Date)) " +
-                              "and level = (Select level From pulls.users Where user_id = " + userId + ")::text;";
-            int sumPullsFromProgram = conn.ExecuteScalar<int>(sqlQuery);
+            sqlQuery = @"Select sum(pulls) From pulls.lvl_user_program WHERE " +
+                              "week = (Select week From pulls.day_result WHERE date = CAST(@date as Date)) " +
+                              "and level = (Select level From pulls.users Where user_id = @users.user_id)::text;";
+            int sumPullsFromProgram = conn.ExecuteScalar<int>(sqlQuery, new { @user_id = userId, @date = date });
 
             if (pulls < sumPullsFromProgram && sumPullsFromProgram > 0)
             {
@@ -150,8 +136,13 @@ namespace PullUpsDapper
             {
                 checkResult = "выполнил";
             }
-            sqlQuery = "UPDATE pulls.day_result Set pulls = '" + pulls + "' WHERE day_result.user_id = " + userId + " and date = CAST('" + date + "' as Date);";
-            conn.Execute(sqlQuery);
+            //sqlQuery = "UPDATE pulls.day_result Set pulls = '" + pulls + "' WHERE day_result.user_id = " + userId + " and date = CAST('" + date + "' as Date);";
+            //conn.Execute(sqlQuery);
+
+            sqlQuery = @"UPDATE pulls.day_result Set pulls = @pulls WHERE day_result.user_id = @user_id and day_result.date = CAST(@date as Date);";
+            conn.Execute(sqlQuery, new { @pulls = pulls, @user_id = userId, @date = date });
+
+
             UserDayProgram.DayReport = false;
             conn.Close();
 
@@ -169,13 +160,17 @@ namespace PullUpsDapper
             for (int i = 0; i < result.Count; i++)
             {
                 LevelProgram res = result[i];
-                var sqlQuery = "INSERT INTO pulls.lvl_user_program (level, week, approach, pulls)  VALUES ('"
-                    + res.Level + "', '"
-                    + res.Week + "', '"
-                    + res.Approach + "', '"
-                    + res.Pulls + "'"
-                    + ")";
-                conn.Execute(sqlQuery);
+                //var sqlQuery = "INSERT INTO pulls.lvl_user_program (level, week, approach, pulls)  VALUES ('"
+                //    + res.Level + "', '"
+                //    + res.Week + "', '"
+                //    + res.Approach + "', '"
+                //    + res.Pulls + "'"
+                //    + ")";
+                //conn.Execute(sqlQuery);
+                var sqlQuery = @"INSERT INTO pulls.lvl_user_program (level, week, approach, pulls) VALUES (@level, @week, @approach, @pulls)";
+                conn.Execute(sqlQuery, new { @level = res.Level, @week = res.Week, @approach = res.Approach, @pulls = res.Pulls });
+
+
             }
             conn.Close();
         }
@@ -183,11 +178,9 @@ namespace PullUpsDapper
         {
             ConnString = DBConnection.ConnectionString();
             using var conn = new NpgsqlConnection(ConnString);
-            conn.Execute("DELETE From pulls.day_result Where day_result.user_id = " + userId + ";");
-
-            conn.Execute("UPDATE pulls.users SET level = null Where users.user_id = " + userId + ";");
+            conn.Execute(@"DELETE From pulls.day_result Where day_result.user_id = @user_id;", new { user_id = userId });
+            conn.Execute(@"UPDATE pulls.users SET level = null Where users.user_id = @user_id;", new { user_id = userId });
             conn.Close();
-
         }
     }
 }
